+++ src/main/java/org/schakalacka/java/raytracing/renderers/RenderWorldWithPlanes.java
@@ -1,9 +1,9 @@
 package org.schakalacka.java.raytracing.renderers;
 
 import org.schakalacka.java.raytracing.PPMExporter;
-import org.schakalacka.java.raytracing.geometry.algebra.MatrixProvider;
-import org.schakalacka.java.raytracing.geometry.algebra.Tuple;
-import org.schakalacka.java.raytracing.geometry.algebra.UVMapping;
+import org.schakalacka.java.raytracing.math.MatrixProvider;
+import org.schakalacka.java.raytracing.math.Tuple;
+import org.schakalacka.java.raytracing.math.UVMapping;
 import org.schakalacka.java.raytracing.geometry.objects.Plane;
 import org.schakalacka.java.raytracing.geometry.objects.Sphere;
 import org.schakalacka.java.raytracing.geometry.patterns.CheckerPattern;
@@ -25,27 +25,27 @@ public class RenderWorldWithPlanes {
 
         var floor = new Plane();
         CheckerPattern floorPattern = new CheckerPattern(new Color(0.3,0.3,0.3), Color.WHITE);
-        floorPattern.setTransformationMatrix(MatrixProvider.rotationY(Math.toRadians(45)));
+        floorPattern.setTransformationMatrix(MatrixProvider.rotationY((float) Math.toRadians(45)));
 
         floor.setMaterial(Material.newMaterial().color(new Color(1, 0.9, 0.9)).ambient(0.5).pattern(floorPattern).specular(0).create());
 
         var backDrop1 = new Plane();
         backDrop1.setTransformationMatrix(MatrixProvider.translation(0, 0, 5)
-                .mulM(MatrixProvider.rotationY(-Math.PI / 4))
-                .mulM(MatrixProvider.rotationX(Math.PI / 2))
+                .mulM(MatrixProvider.rotationY((float) (-Math.PI / 4)))
+                .mulM(MatrixProvider.rotationX((float) (Math.PI / 2)))
         );
         backDrop1.setMaterial(Material.newMaterial().color(new Color(1, 0.9, 0.9)).ambient(0.5).diffuse(0.8).create());
 
         var backDrop2 = new Plane();
         backDrop2.setTransformationMatrix(MatrixProvider.translation(0, 0, 5)
-                .mulM(MatrixProvider.rotationY(Math.PI / 4))
-                .mulM(MatrixProvider.rotationX(Math.PI / 2))
+                .mulM(MatrixProvider.rotationY((float) (Math.PI / 4)))
+                .mulM(MatrixProvider.rotationX((float) (Math.PI / 2)))
         );
 
         backDrop2.setMaterial(Material.newMaterial().color(new Color(1, 0.9, 0.9)).ambient(0.5).diffuse(0.8).create());
 
         var middleSphere = new Sphere();
-        middleSphere.setTransformationMatrix(MatrixProvider.translation(-0.5, 1, 0.5));
+        middleSphere.setTransformationMatrix(MatrixProvider.translation(-0.5f, 1, 0.5f));
         middleSphere.setMaterial(Material.newMaterial()
                 .color(new Color(0.1, 1, 0.5))
                 .diffuse(0.7)
@@ -54,7 +54,7 @@ public class RenderWorldWithPlanes {
                 .create());
 
         var rightSphere = new Sphere();
-        rightSphere.setTransformationMatrix(MatrixProvider.translation(1.5, 0.5, -0.5).mulM(MatrixProvider.scaling(0.5, 0.5, 0.5)));
+        rightSphere.setTransformationMatrix(MatrixProvider.translation(1.5f, 0.5f, -0.5f).mulM(MatrixProvider.scaling(0.5f, 0.5f, 0.5f)));
         rightSphere.setMaterial(Material.newMaterial()
                 .color(new Color(0.5, 1, 0.1))
                 .diffuse(0.7)
@@ -62,7 +62,7 @@ public class RenderWorldWithPlanes {
                 .create());
 
         var leftSphere = new Sphere();
-        leftSphere.setTransformationMatrix(MatrixProvider.translation(-1.5, 0.33, -0.75).mulM(MatrixProvider.scaling(0.33, 0.33, 0.33)));
+        leftSphere.setTransformationMatrix(MatrixProvider.translation(-1.5f, 0.33f, -0.75f).mulM(MatrixProvider.scaling(0.33f, 0.33f, 0.33f)));
         leftSphere.setMaterial(Material.newMaterial()
                 .color(new Color(1, 0.2, 0.1))
                 .diffuse(0.7)
@@ -79,7 +79,7 @@ public class RenderWorldWithPlanes {
         var camera = new Camera(width, height, Math.PI / 3);
         camera.setTransformationMatrix(ViewTransformation
                 .transform(
-                        Tuple.point(0, 1.5, -5),
+                        Tuple.point(0, 1.5f, -5),
                         Tuple.point(0, 1, 0),
                         Tuple.vector(0, 1, 0)
                 )
